Refactored advanced entropy operation into entropy operation
